refactor(encode): name the stream seam and take the cfg walls down
Review of the fdk backend found the dispatch repeating a per-feature match in
three places and the offline routes doing the same across two, which is what
tipped both files past the cfg-density gate. `AacStream` names what the two
backends already agreed on, so the dispatch resolves once at open and the
per-codec routing lives in its own file next to the byte path.

The f32 the caller pushes now reaches i16 through `num-traits` and rounds to
the nearest sample, which is what the cast group's policy asks for and drops
the lint suppression the first cut carried. A short frame from libfdk is a
typed backend error rather than a silent stall, the frame buffer moves on a
cursor instead of shifting per frame, and `create_packaged` answers for the
codec it was asked about, as its documentation always said it did.
