+++ crates/kithara-encode/Cargo.toml
@@ -30,6 +30,7 @@ tls-rustls = ["kithara-stream/tls-rustls"]
 [dependencies]
 kithara-stream = { workspace = true }
 
+num-traits = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
