+++ crates/kithara-encode/src/offline/mod.rs
@@ -1,0 +1,6 @@
+mod bytes;
+mod packaged;
+#[cfg(all(test, not(feature = "ffmpeg")))]
+mod tests;
+
+pub(crate) use self::packaged::OfflineEncoder;
