+++ crates/kithara-encode/src/offline/packaged.rs
@@ -10,24 +10,14 @@ use crate::{
 };
 
 /// Offline encoding over a finite PCM source: complete encoded bytes and
-/// packaged access units. Each codec is routed to the backend that owns it.
+/// packaged access units. Each codec is routed to the backend that owns it, and
+/// one whose backend this build does not carry is unsupported.
 #[derive(Debug, Clone, Copy)]
 pub(crate) struct OfflineEncoder;
 
 impl InnerEncoder for OfflineEncoder {
     fn encode_bytes(&self, request: BytesEncodeRequest<'_>) -> EncodeResult<EncodedBytes> {
-        #[cfg(feature = "ffmpeg")]
-        {
-            crate::ffmpeg::bytes::encode_bytes_audio(&request)
-        }
-
-        #[cfg(not(feature = "ffmpeg"))]
-        {
-            let _ = request;
-            Err(EncodeError::InvalidInput(
-                "byte encoding needs the `ffmpeg` feature".to_owned(),
-            ))
-        }
+        super::bytes::encode(&request)
     }
 
     fn encode_packaged(&self, request: PackagedEncodeRequest<'_>) -> EncodeResult<EncodedTrack> {
@@ -58,42 +48,5 @@ impl InnerEncoder for OfflineEncoder {
             AudioCodec::Flac => Ok(FlacFFmpegEncoder::frame_samples()),
             codec => Err(EncodeError::UnsupportedCodec(codec)),
         }
-    }
-}
-
-/// A build without `FFmpeg` still answers every offline route, and answers the
-/// ones it cannot serve with an error rather than a panic.
-#[cfg(all(test, not(feature = "ffmpeg")))]
-mod tests {
-    use kithara_stream::AudioCodec;
-
-    use super::OfflineEncoder;
-    use crate::{
-        BytesEncodeRequest, BytesEncodeTarget, EncodeError, InnerEncoder, test_pcm::TestPcm,
-    };
-
-    #[test]
-    fn byte_encoding_reports_the_missing_backend() {
-        let pcm = TestPcm::sawtooth(1_024, 48_000, 2);
-
-        let error = OfflineEncoder
-            .encode_bytes(BytesEncodeRequest {
-                pcm: &pcm,
-                target: BytesEncodeTarget::Mp3,
-                bit_rate: None,
-            })
-            .map(|_| ())
-            .expect_err("no FFmpeg, no byte encoding");
-
-        assert!(matches!(error, EncodeError::InvalidInput(_)), "{error}");
-    }
-
-    #[test]
-    fn a_codec_whose_backend_is_absent_is_unsupported() {
-        let error = OfflineEncoder
-            .packaged_frame_samples(AudioCodec::Flac)
-            .expect_err("no FFmpeg, no FLAC");
-
-        assert!(matches!(error, EncodeError::UnsupportedCodec(_)), "{error}");
     }
 }
