+++ crates/kithara-encode/src/stream/mod.rs
@@ -1,0 +1,6 @@
+mod encoder;
+#[cfg(test)]
+mod tests;
+
+pub(crate) use self::encoder::{AacStream, StreamParams};
+pub use self::encoder::{StreamBackend, StreamEncoder};
